politeiawww: Move sessions into legacy package.
The existing sessions store is tangled up with the legacy user database
and politeia specific session values, so it's being moved into the
legacy package. The implementation does not have proper separation of
concerns and will not be usable by the new plugin model.
