+++ politeiawww/legacy/politeiawww.go
@@ -34,6 +34,7 @@ import (
 	ghtracker "github.com/decred/politeia/politeiawww/legacy/codetracker/github"
 	"github.com/decred/politeia/politeiawww/legacy/comments"
 	"github.com/decred/politeia/politeiawww/legacy/pi"
+	"github.com/decred/politeia/politeiawww/legacy/sessions"
 	"github.com/decred/politeia/politeiawww/legacy/ticketvote"
 	"github.com/decred/politeia/politeiawww/legacy/user"
 	"github.com/decred/politeia/politeiawww/legacy/user/cockroachdb"
@@ -41,7 +42,6 @@ import (
 	"github.com/decred/politeia/politeiawww/legacy/user/mysql"
 	"github.com/decred/politeia/politeiawww/mail"
 	"github.com/decred/politeia/politeiawww/records"
-	"github.com/decred/politeia/politeiawww/sessions"
 	"github.com/decred/politeia/politeiawww/websockets"
 	"github.com/decred/politeia/util"
 	"github.com/decred/politeia/wsdcrdata"
