+++ politeiawww/legacy/sessions/store.go
