+++ politeiawww/records/records.go
@@ -12,8 +12,8 @@ import (
 	v1 "github.com/decred/politeia/politeiawww/api/records/v1"
 	"github.com/decred/politeia/politeiawww/config"
 	"github.com/decred/politeia/politeiawww/events"
+	"github.com/decred/politeia/politeiawww/legacy/sessions"
 	"github.com/decred/politeia/politeiawww/legacy/user"
-	"github.com/decred/politeia/politeiawww/sessions"
 	"github.com/decred/politeia/util"
 )
